rename and enhanced userStore
