+++ src/components/admin/AdminNavbar.vue
@@ -12,7 +12,7 @@ defineEmits(["toggle-sidebar"]);
 
 <script>
 
-import { useUserStore } from "@/store/user";
+import { useUserStore } from "@/store/userStore";
 import router from "@/router";
 
 export default {
